requirements.txt
protobuf
bitstruct
attrs
pyzmq
